Don't cause the mongodb plugin to fail
This was visible when there were services with transactional = 'mongo' set.
It seems that the mongo plugin was seeing our ProxyFactoryBean-wrapped
services and assumed that they were wrapped in a TransactionalProxyFactoryBean,
which they weren't. The fix is to make sure that we only do our stuff after
the mongodb plugin has loaded.

There may potentially be other cases like this, but I don't have time to hunt
them down, let's see if there are any other error reports.